Admin and Users implemented
Admin and Users implemented

+++ backend/src/main/java/com/example/techrental/model/User.java
@@ -1,107 +1,3 @@
-// package com.example.techrental.model;
-
-// import jakarta.persistence.*;
-// import java.util.Date;
-
-// @Entity
-// @Table(name = "users")
-// public class User {
-
-//     @Id
-//     @GeneratedValue(strategy = GenerationType.IDENTITY)
-//     private Long id;
-
-//     private String password;
-//     private String role;
-
-//     private String firstName;
-//     private String lastName;
-
-//     @Column(unique = true)
-//     private String email;
-//     private String phoneNumber;
-    
-
-//     @Temporal(TemporalType.TIMESTAMP)
-//     private Date registrationDate;
-
-//     @PrePersist
-//     protected void onCreate() {
-//         registrationDate = new Date();
-//     }
-
-
-//     // Getters and Setters
-
-//     public Long getId() {
-//         return id;
-//     }
-
-//     public void setId(Long id) {
-//         this.id = id;
-//     }
-
-//     public String getPassword() {
-//         return password;
-//     }
-
-//     public void setPassword(String password) {
-//         this.password = password;
-//     }
-
-//     public String getRole() {
-//         return role;
-//     }
-
-//     public void setRole(String role) {
-//         this.role = role;
-//     }
-
-//     public String getFirstName() {
-//         return firstName;
-//     }
-
-//     public void setFirstName(String firstName) {
-//         this.firstName = firstName;
-//     }
-
-//     public String getLastName() {
-//         return lastName;
-//     }
-
-//     public void setLastName(String lastName) {
-//         this.lastName = lastName;
-//     }
-
-//     public String getEmail() {
-//         return email;
-//     }
-
-//     public void setEmail(String email) {
-//         this.email = email;
-//     }
-
-//     public String getPhoneNumber() {
-//         return phoneNumber;
-//     }
-
-//     public void setPhoneNumber(String phoneNumber) {
-//         this.phoneNumber = phoneNumber;
-//     }
-
-//     public Date getRegistrationDate() {
-//         return registrationDate;
-//     }
-
-//     public void setRegistrationDate(Date registrationDate) {
-//         this.registrationDate = registrationDate;
-//     }
-// }
-
-
-
-
-
 package com.example.techrental.model;
 
 import jakarta.persistence.*;
@@ -130,7 +26,7 @@ public class User {
 
     @Column(unique = true)
     @NotNull(message = "Phone number must not be null")
-    private int phoneNumber;
+    private long phoneNumber;
 
 
     @Temporal(TemporalType.TIMESTAMP)
@@ -192,11 +88,11 @@ public class User {
         this.email = email;
     }
 
-    public int getPhoneNumber() {
+    public long getPhoneNumber() {
         return phoneNumber;
     }
 
-    public void setPhoneNumber(int phoneNumber) {
+    public void setPhoneNumber(long phoneNumber) {
         this.phoneNumber = phoneNumber;
     }
