+++ backend/src/main/java/com/example/techrental/service/UserService.java
@@ -1,22 +1,3 @@
-// package com.example.techrental.service;
-
-// import com.example.techrental.model.User;
-// import java.util.List;
-
-// public interface UserService {
-//     void saveUser(User user);
-//     User findByEmail(String email);
-//     User findById(Long id);
-//     List<User> findAllUsersSortedById();
-//     List<User> searchUsers(String query);
-//     void deleteUserById(Long id);
-   
-// }
-
-
-
-
-
 package com.example.techrental.service;
 
 import com.example.techrental.model.User;
@@ -31,7 +12,7 @@ public interface UserService {
     void deleteUserById(Long id);
 
     boolean existsByEmail(String email);
-    boolean existsByPhoneNumber(int phoneNumber);
+    boolean existsByPhoneNumber(long phoneNumber);
 
 }
