+++ backend/src/main/java/com/example/techrental/service/UserServiceImpl.java
@@ -1,56 +1,3 @@
-// package com.example.techrental.service;
-
-// import com.example.techrental.model.User;
-// import com.example.techrental.repository.UserRepository;
-// import org.springframework.beans.factory.annotation.Autowired;
-// import org.springframework.stereotype.Service;
-
-// import java.util.List;
-
-// @Service
-// public class UserServiceImpl implements UserService {
-
-//     @Autowired
-//     private UserRepository userRepository;
-
-//     @Override
-//     public void saveUser(User user) {
-//         userRepository.save(user);
-//     }
-
-//     @Override
-//     public User findByEmail(String email) {
-//         return userRepository.findByEmail(email);
-//     }
-
-//     @Override
-//     public User findById(Long id) {
-//         return userRepository.findById(id).orElse(null); 
-//     }
-
-//     @Override
-//     public List<User> findAllUsersSortedById() {
-//         return userRepository.findAllByOrderByIdAsc();
-//     }
-
-//     @Override
-//     public List<User> searchUsers(String query) {
-//         return userRepository.findByFirstNameContainingIgnoreCaseOrLastNameContainingIgnoreCaseOrEmailContainingIgnoreCase(query, query, query);
-//     }
-
-//     @Override
-//     public void deleteUserById(Long id) {
-//         userRepository.deleteById(id);
-//     }
-
-// }
-
-
-
-
-
-
-
 package com.example.techrental.service;
 
 import com.example.techrental.model.User;
@@ -102,9 +49,10 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public boolean existsByPhoneNumber(int phoneNumber) {
+    public boolean existsByPhoneNumber(long phoneNumber) {
         return userRepository.existsByPhoneNumber(phoneNumber);
     }
+
 
 }
